backend/requirements.txt
# Backend Python Worker
pydantic>=2.0
# docling>=2.0  # Uncomment when ready for real Docling integration
PyMuPDF  # Relaxed version constraint for Python 3.14
Pillow==10.1.0
